Initial Build of Updated Architecture
