+++ D4e/D4e_Utilities.py
@@ -1,34 +1,24 @@
 # imports
 import asyncio
-import datetime
-import inspect
 import logging
-import os
-import sys
 
 import d20
 import discord
 from discord import Interaction
-
-from dotenv import load_dotenv
-from sqlalchemy import select, false
 from sqlalchemy.exc import NoResultFound
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import sessionmaker
+
+from Base.Tracker import get_init_list
+from Base.Utilities import Utilities
+from database_models import get_tracker, get_condition
+from error_handling_reporting import error_not_initialized, ErrorReport
+from utils.Char_Getter import get_character
 from utils.Tracker_Getter import get_tracker_model
+from utils.utils import get_guild
+
 
 # import D4e.d4e_functions
-import PF2e.pf2_functions
-import time_keeping_functions
-
-from database_models import get_tracker, get_condition, get_macro
-from error_handling_reporting import error_not_initialized, ErrorReport
-# from initiative import get_guild, PF2AddCharacterModal, D4eAddCharacterModal, update_pinned_tracker
-from utils.Char_Getter import get_character
-from utils.utils import get_guild
-# from utils.Tracker_Getter import get_tracker_model
-from Base.Generic_Utilities import Utilities
-from Base.Tracker import get_init_list
 
 class D4e_Utilities(Utilities):
     def __init__(self, ctx, guild, engine):
