+++ D4e/d4e_functions.py
@@ -1,24 +1,9 @@
 # pf2_functions.py
 
 
-import logging
-
 # imports
 
-import discord
-from sqlalchemy import or_, select
-from sqlalchemy.exc import NoResultFound
-from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy.orm import sessionmaker
-
 import d20
-
-from utils.utils import get_guild
-from utils.Char_Getter import get_character
-
-from database_models import Global, get_condition, get_tracker
-from error_handling_reporting import ErrorReport, error_not_initialized
-from utils.parsing import ParseModifiers
 
 D4e_attributes = ["AC", "Fort", "Reflex", "Will"]
 D4e_base_roll = d20.roll(f"{10}")
@@ -38,34 +23,5 @@ def D4e_eval_success(dice_result: d20.RollResult, goal: d20.RollResult):
 
 
 # Builds the tracker string. Updated to work with block initiative
-
-
-# Checks to see if the user of the slash command is the GM, returns a boolean
-async def gm_check(ctx, engine, guild=None):
-    if ctx is None and guild is None:
-        raise LookupError("No guild reference")
-
-    async_session = sessionmaker(engine, expire_on_commit=False, class_=AsyncSession)
-    try:
-        async with async_session() as session:
-            if ctx is None:
-                result = await session.execute(select(Global).where(Global.id == guild.id))
-            else:
-                result = await session.execute(
-                    select(Global).where(
-                        or_(
-                            Global.tracker_channel == ctx.interaction.channel_id,
-                            Global.gm_tracker_channel == ctx.interaction.channel_id,
-                        )
-                    )
-                )
-            guild = result.scalars().one()
-
-            if int(guild.gm) != int(ctx.interaction.user.id):
-                return False
-            else:
-                return True
-    except Exception:
-        return False
