+++ database_models.py
@@ -1,11 +1,9 @@
 # database_models.py
 
-import os
 import logging
 
 import discord
 import sqlalchemy as db
-from dotenv import load_dotenv
 from sqlalchemy import Column, JSON
 from sqlalchemy import ForeignKey
 from sqlalchemy import Integer, BigInteger
@@ -14,29 +12,8 @@ from sqlalchemy import or_, select
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import declarative_base
 from sqlalchemy.orm import sessionmaker
-# from initiative import get_guild
 
-# define global variables
-# import initiative
-# import initiative
 from EPF import EPF_Support
-
-load_dotenv(verbose=True)
-if os.environ["PRODUCTION"] == "True":
-    TOKEN = os.getenv("TOKEN")
-    USERNAME = os.getenv("Username")
-    PASSWORD = os.getenv("Password")
-    HOSTNAME = os.getenv("Hostname")
-    PORT = os.getenv("PGPort")
-else:
-    TOKEN = os.getenv("BETA_TOKEN")
-    USERNAME = os.getenv("BETA_Username")
-    PASSWORD = os.getenv("BETA_Password")
-    HOSTNAME = os.getenv("BETA_Hostname")
-    PORT = os.getenv("BETA_PGPort")
-
-GUILD = os.getenv("GUILD")
-SERVER_DATA = os.getenv("SERVERDATA")
 
 Base = declarative_base()
