+++ dice_roller_cog.py
@@ -2,20 +2,14 @@
 import logging
 import os
 
+import d20
 # imports
 import discord
 from discord import option
 from discord.ext import commands
 from dotenv import load_dotenv
-from sqlalchemy import select, or_
-from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy.orm import sessionmaker
-
-import d20
 
 # import initiative
-from database_models import Global
-from database_operations import get_asyncio_db_engine
 from error_handling_reporting import ErrorReport
 from utils.parsing import opposed_roll
 from utils.utils import get_guild
